refactor: AgeDiscountPolicy가 계산을 수행하도록 변경
