+++ src/main/java/subway/domain/fare/FareCalculator.java
@@ -8,14 +8,13 @@ public class FareCalculator {
     private final FarePolicy farePolicy;
     private final DiscountPolicy discountPolicy;
 
-    public FareCalculator(final FarePolicy farePolicy,final DiscountPolicy discountPolicy) {
+    public FareCalculator(final FarePolicy farePolicy, final DiscountPolicy discountPolicy) {
         this.farePolicy = farePolicy;
         this.discountPolicy = discountPolicy;
     }
 
     public int calculate(final FareInformation fareInformation) {
-        int fare = 0;
-        fare += farePolicy.calculate(fareInformation);
+        int fare = farePolicy.calculate(fareInformation);
         return discountPolicy.calculate(fare, fareInformation);
     }
 }
